requirements.txt
Flask==2.0.1
pymongo==4.3.3
Flask-Cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
